@@ -2,10 +2,7 @@ namespace CloudRepublic.BenchMark.Domain.Enums
 {
     public enum Runtime
     {
-        Csharp = 0,
-        Nodejs = 1,
-        Python = 2,
-        Java = 3,
-        Fsharp = 4,
+        FunctionsV4 = 0,
+        Firebase = 1,
     }
 }
